+++ app/src/main/java/edu/uiuc/cs427app/HomeActivity.java
@@ -2,13 +2,74 @@ package edu.uiuc.cs427app;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import androidx.navigation.ui.AppBarConfiguration;
+import edu.uiuc.cs427app.databinding.ActivityHomeBinding;
 
-public class HomeActivity extends AppCompatActivity {
 
+public class HomeActivity extends AppCompatActivity implements View.OnClickListener {
+
+    private AppBarConfiguration appBarConfiguration;
+    private ActivityHomeBinding binding;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_home);
+
+        // Initializing the UI components
+        // The list of locations should be customized per user (change the implementation so that
+        // buttons are added to layout programmatically
+        Button buttonChampaign = findViewById(R.id.buttonChampaign);
+        Button buttonChicago = findViewById(R.id.buttonChicago);
+        Button buttonLA = findViewById(R.id.buttonLA);
+        Button buttonNew = findViewById(R.id.buttonAddLocation);
+
+        buttonChampaign.setOnClickListener(this);
+        buttonChicago.setOnClickListener(this);
+        buttonLA.setOnClickListener(this);
+        buttonNew.setOnClickListener(this);
+
+        Button buttonLogout = findViewById(R.id.btnLogout);
+        buttonLogout.setOnClickListener(this);
     }
+
+    @Override
+    public void onClick(View view) {
+        Intent intent;
+        switch (view.getId()) {
+            case R.id.buttonChampaign:
+                intent = new Intent(this, DetailsActivity.class);
+                intent.putExtra("city", "Champaign");
+                startActivity(intent);
+                break;
+            case R.id.buttonChicago:
+                intent = new Intent(this, DetailsActivity.class);
+                intent.putExtra("city", "Chicago");
+                startActivity(intent);
+                break;
+            case R.id.buttonLA:
+                intent = new Intent(this, DetailsActivity.class);
+                intent.putExtra("city", "Los Angeles");
+                startActivity(intent);
+                break;
+            case R.id.btnLogout:
+                intent = new Intent(this, LoginActivity.class);
+                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                startActivity(intent);
+                break;
+            case R.id.buttonAddLocation:
+                // Implement this action to add a new location to the list of locations
+                break;
+        }
+    }
+
+//    public void logout(View view) {
+//        Intent intent = new Intent(this, LoginActivity.class);
+//        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+//        startActivity(intent);
+//        finish();
+//    }
 }
