+++ app/src/main/res/layout/activity_home.xml
@@ -1,59 +1,52 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <LinearLayout
-        android:id="@+id/details"
-        android:layout_width="409dp"
-        android:layout_height="601dp"
-        android:orientation="vertical"
-        tools:layout_editor_absoluteX="1dp"
-        tools:layout_editor_absoluteY="1dp"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" tools:context=".HomeActivity">
+    <LinearLayout android:id="@+id/details" android:layout_width="409dp"
+        android:layout_height="601dp" android:orientation="vertical"
+        tools:layout_editor_absoluteX="1dp" tools:layout_editor_absoluteY="1dp"
         tools:ignore="MissingConstraints">
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+        <RelativeLayout android:layout_width="match_parent" android:layout_height="wrap_content"
             android:padding="16dp">
-
-            <TextView
-                android:id="@+id/welcomeText"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="DashBoard"
-                android:textAlignment="center"
-                android:textColor="@color/black"
-                android:textSize="30dp"
-                android:textStyle="bold" />
-
-            <Button
-                android:id="@+id/btnLogout"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Logout"
-                android:layout_alignParentEnd="true"
-                android:layout_centerVertical="true"
-                android:onClick="logoutMenu"/>
+            <TextView android:id="@+id/welcomeText" android:layout_width="wrap_content"
+                android:layout_height="wrap_content" android:text="DashBoard"
+                android:textAlignment="center" android:textColor="@color/black"
+                android:textSize="30dp" android:textStyle="bold" />
+            <Button android:id="@+id/btnLogout" android:layout_width="wrap_content"
+                android:layout_height="wrap_content" android:text="Logout"
+                android:layout_alignParentEnd="true" android:layout_centerVertical="true" />
         </RelativeLayout>
-
-
-
-        <TextView
-            android:id="@+id/cityInfo"
-            android:layout_width="match_parent"
-            android:layout_height="268dp" />
-
-        <Button
-            android:id="@+id/mapButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Show City" />
-
-
-
+        <LinearLayout android:layout_width="match_parent" android:layout_height="44dp"
+            android:orientation="horizontal">
+            <TextView android:id="@+id/textView10" android:layout_width="wrap_content"
+                android:layout_height="wrap_content" android:layout_weight="1"
+                android:text="Champaign" />
+            <Button android:id="@+id/buttonChampaign" android:layout_width="45dp"
+                android:layout_height="wrap_content" android:layout_weight="1"
+                android:text="Show Details" />
+        </LinearLayout>
+        <LinearLayout android:layout_width="match_parent" android:layout_height="45dp"
+            android:orientation="horizontal">
+            <TextView android:id="@+id/textView12" android:layout_width="wrap_content"
+                android:layout_height="wrap_content" android:layout_weight="1"
+                android:text="Chicago" />
+            <Button android:id="@+id/buttonChicago" android:layout_width="25dp"
+                android:layout_height="wrap_content" android:layout_weight="1"
+                android:text="Show Details" />
+        </LinearLayout>
+        <LinearLayout android:layout_width="match_parent" android:layout_height="39dp"
+            android:orientation="horizontal">
+            <TextView android:id="@+id/textView13" android:layout_width="wrap_content"
+                android:layout_height="wrap_content" android:layout_weight="1"
+                android:text="Los Angeles" />
+            <Button android:id="@+id/buttonLA" android:layout_width="52dp"
+                android:layout_height="wrap_content" android:layout_weight="1"
+                android:text="Show Details" />
+        </LinearLayout>
+        <TextView android:id="@+id/textView14" android:layout_width="match_parent"
+            android:layout_height="156dp" />
+        <Button android:id="@+id/buttonAddLocation" android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:text="Add a location" />
     </LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
