+++ app/src/main/res/layout/activity_main.xml
@@ -1,105 +1,16 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <LinearLayout
-        android:layout_width="409dp"
-        android:layout_height="729dp"
-        android:orientation="vertical"
-        tools:layout_editor_absoluteX="1dp"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" tools:context=".MainActivity">
+    <LinearLayout android:layout_width="409dp" android:layout_height="729dp"
+        android:orientation="vertical" tools:layout_editor_absoluteX="1dp"
         tools:layout_editor_absoluteY="1dp">
-
-        <TextView
-            android:id="@+id/textView3"
-            android:layout_width="match_parent"
-            android:layout_height="50dp"
-            android:text="CS427 Project App" />
-
-        <TextView
-            android:id="@+id/textView4"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="List of Locations" />
-
-        <TextView
-            android:id="@+id/textView5"
-            android:layout_width="match_parent"
+        <TextView android:id="@+id/textView3" android:layout_width="match_parent"
+            android:layout_height="50dp" android:text="CS427 Project App" />
+        <TextView android:id="@+id/textView4" android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:text="List of Locations" />
+        <TextView android:id="@+id/textView5" android:layout_width="match_parent"
             android:layout_height="wrap_content" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="44dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:id="@+id/textView10"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="Champaign" />
-
-            <Button
-                android:id="@+id/buttonChampaign"
-                android:layout_width="45dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="Show Details" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="45dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:id="@+id/textView12"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="Chicago" />
-
-            <Button
-                android:id="@+id/buttonChicago"
-                android:layout_width="25dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="Show Details" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="39dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:id="@+id/textView13"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="Los Angeles" />
-
-            <Button
-                android:id="@+id/buttonLA"
-                android:layout_width="52dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="Show Details" />
-        </LinearLayout>
-
-        <TextView
-            android:id="@+id/textView14"
-            android:layout_width="match_parent"
-            android:layout_height="156dp" />
-
-        <Button
-            android:id="@+id/buttonAddLocation"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Add a location" />
-
     </LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
